generifies event handler testing code and moves it to common helper file
